requirements.txt
# Agent Orchestrator - Polymarket
httpx>=0.27.0
anthropic>=0.39.0
tavily-python>=0.5.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
